refacto alchemy, thegraph and covalent apis
